Refactor state management: move state_manager to a new file and update imports
